restructure gui code
