+++ gui/__init__.py
@@ -1,0 +1,3 @@
+from .gui import MainDialog
+from .config import ConfigDialog
+from .tools import UploadThread, QtHandler
