+++ requirements.txt
@@ -9,4 +9,5 @@ cerberus
 pynsist==2.3
 argparse
 selenium
-PyQt5==5.12.2
+PyQt5==5.12.1
+PyQt5-stubs==5.12.1
