+++ windows-gui-installer.cfg
@@ -1,7 +1,7 @@
 [Application]
 name=IRIDA Sequence Uploader GUI
-version=1.0
-entry_point=gui:main
+version=0.3
+entry_point=upload_gui:main
 icon=images/icon.ico
 # Uncomment this to have a console show alongside the application
 # console=true
@@ -17,11 +17,13 @@ packages=upload_run
     config
     core
     global_settings
+    gui
     messaging
     model
     parsers
     progress
     rauth
+    pkg_resources
     cerberus
     appdirs
     requests
@@ -31,9 +33,9 @@ packages=upload_run
     idna
     argparse
 
-pypi_wheels= PyQt5==5.9.2
-    sip==4.19.8
+pypi_wheels= PyQt5==5.12.1
+    PyQt5-sip==4.19.17
 
 # Adds iridagui as a command to cmd
 [Command iridagui]
-entry_point=gui:main
+entry_point=upload_gui:main
